Update to include Mirror Check
Removes a single solution. If you check all solutions at the same time,
they create a similar pattern at a single step. When doing this, the
board mirrors itself at certain steps, causing there to be a significant
deprecation of solutions. From 17 to 7, Solving for each peg seperately
causes a single loss of a single solution which therefore must have been
a mirror solution that was not caught earlier by previous checks.